+++ src/main/java/de/theidler/create_mobile_packages/index/CMPMenuTypes.java
@@ -6,8 +6,12 @@ import de.theidler.create_mobile_packages.CreateMobilePackages;
 import de.theidler.create_mobile_packages.blocks.bee_port.BeePortMenu;
 import de.theidler.create_mobile_packages.blocks.bee_port.BeePortScreen;
 import de.theidler.create_mobile_packages.items.mobile_packager.*;
+import de.theidler.create_mobile_packages.items.portable_stock_ticker.PortableStockTicker;
 import de.theidler.create_mobile_packages.items.portable_stock_ticker.PortableStockTickerMenu;
 import de.theidler.create_mobile_packages.items.portable_stock_ticker.PortableStockTickerScreen;
+import de.theidler.create_mobile_packages.items.portable_stock_ticker.trash_menu.TrashMenu;
+import de.theidler.create_mobile_packages.items.portable_stock_ticker.trash_menu.TrashScreen;
+import net.minecraft.world.item.ItemStack;
 
 public class CMPMenuTypes {
 
@@ -37,6 +41,19 @@ public class CMPMenuTypes {
                     "mobile_packager_edit_menu",
                     MobilePackagerEditMenu::new,
                     () -> MobilePackagerEditScreen::new
+            ).register();
+
+    public static final MenuEntry<TrashMenu> TRASH_MENU =
+            CreateMobilePackages.REGISTRATE.menu(
+                    "trash_menu",
+                    (trashMenuType, containerId, playerInventory) -> {
+                        ItemStack pstStack = PortableStockTicker.find(playerInventory);
+                        PortableStockTicker pst = (pstStack != null && pstStack.getItem() instanceof PortableStockTicker)
+                                ? (PortableStockTicker) pstStack.getItem()
+                                : null;
+                        return new TrashMenu(containerId, playerInventory, pst);
+                    },
+                    () -> TrashScreen::new
             ).register();
 
     public static void register() {